+++ Source/POETradeHelper.ItemSearch/ViewModels/AdvancedQuery/IMinMaxFilterViewModel.cs
@@ -4,8 +4,8 @@
     {
         string Text { get; set; }
         bool? IsEnabled { get; set; }
-        int? Min { get; set; }
-        int? Max { get; set; }
+        decimal? Min { get; set; }
+        decimal? Max { get; set; }
         string Current { get; set; }
     }
 }
